% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomes.R
\name{add_genomes}
\alias{add_genomes}
\title{Add genomes to a potato sack}
\usage{
add_genomes(sack, path, validate = TRUE, recursive = FALSE)
}
\arguments{
\item{sack}{PotatoSack object}

\item{path}{Character. Path to protein FASTA (.faa or .fasta) files. Can be:
\itemize{
\item A directory: adds all .faa/.fasta files in directory
\item A wildcard pattern: "~/data/mags/*.faa"
\item A single file: "~/data/genome1.faa"
\item A vector of files
}}

\item{validate}{Logical. If TRUE, validates .faa files (default: TRUE)}

\item{recursive}{Logical. If path is a directory, search recursively (default: FALSE)}
}
\value{
Modified PotatoSack object with genomes added
}
\description{
Register genome files with a potato sack without copying them. Can be called
multiple times to add genomes from different locations.
}
