% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{annotate_genomes_blast}
\alias{annotate_genomes_blast}
\title{Annotate multiple genomes with BLAST}
\usage{
annotate_genomes_blast(genomes, potatoes, config, cleanup = TRUE)
}
\arguments{
\item{genomes}{List of jakomics FILE objects}

\item{potatoes}{List of Potato objects}

\item{config}{Potato config object}

\item{cleanup}{Logical. Remove intermediate files after parsing (default TRUE)}
}
\value{
List of results, one per genome
}
\description{
Annotate multiple genomes with BLAST
}
