% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kofam.R
\name{annotate_genomes_kofam}
\alias{annotate_genomes_kofam}
\title{Annotate multiple genomes with KOfam}
\usage{
annotate_genomes_kofam(genomes, potatoes, config, cleanup = TRUE)
}
\arguments{
\item{genomes}{List of jakomics FILE objects}

\item{potatoes}{List of Potato objects (or single Potato)}

\item{config}{Potato config object}

\item{cleanup}{Logical. Remove intermediate files (.hal, .tsv) after parsing (default TRUE)}
}
\value{
List of results, one per genome
}
\description{
Run KOfam annotation across multiple genomes sequentially.
}
