% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_simple.R
\name{annotate_sack_simple}
\alias{annotate_sack_simple}
\title{Simple annotation - just produce gene-level results}
\usage{
annotate_sack_simple(sack)
}
\arguments{
\item{sack}{PotatoSack object}
}
\description{
Simple annotation - just produce gene-level results
}
