% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato_v2.R
\name{build_graph}
\alias{build_graph}
\title{Build igraph from v2 potato}
\usage{
build_graph(potato_v2, pathway_id = NULL)
}
\arguments{
\item{potato_v2}{Potato loaded with load_potato()}

\item{pathway_id}{Which pathway to build (default: first pathway)}
}
\description{
Build igraph from v2 potato
}
