% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-potato_sack.R
\name{calculate_dir_hash}
\alias{calculate_dir_hash}
\title{Calculate directory hash}
\usage{
calculate_dir_hash(dir_path, pattern = NULL)
}
\arguments{
\item{dir_path}{Path to directory}

\item{pattern}{Optional file pattern to match}
}
\value{
MD5 hash string or NULL
}
\description{
Calculate combined hash of all files in a directory
}
\keyword{internal}
