% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato_hash.R
\name{compute_potato_hash}
\alias{compute_potato_hash}
\title{Compute MD5 hash of potato R object}
\usage{
compute_potato_hash(potato)
}
\arguments{
\item{potato}{Potato S7 object}
}
\value{
Character string with MD5 hash
}
\description{
Computes a hash of the potato's functional content (nodes, edges, scoring, etc.)
to track which version was used for annotation. Excludes metadata fields like
json_path and graph cache that don't affect annotation behavior.
}
\details{
Specifically excludes cosmetic fields that don't affect detection/scoring:
\itemize{
\item x, y, x_compounds, y_compounds (visualization coordinates)
\item notes (documentation only)
}
}
\keyword{internal}
