% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kofam_hal.R
\name{create_kofam_hal}
\alias{create_kofam_hal}
\title{Create .hal file for kofam from potato KO list}
\usage{
create_kofam_hal(sack, potato_names = NULL)
}
\arguments{
\item{sack}{PotatoSack object}

\item{potato_names}{Vector of potato names (NULL = all)}
}
\value{
List with hal_path and modified sack
}
\description{
Create .hal file for kofam from potato KO list
}
