% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{export_potato_dot}
\alias{export_potato_dot}
\title{Export potato graph to graphviz DOT format}
\usage{
export_potato_dot(potato, file)
}
\arguments{
\item{potato}{Potato S7 object}

\item{file}{Output file path}
}
\description{
Export potato graph to graphviz DOT format
}
