% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato_class.R
\name{get_detection_terms}
\alias{get_detection_terms}
\title{Get detection terms for a tool type}
\usage{
get_detection_terms(potato, tool_type)
}
\arguments{
\item{potato}{Potato object}

\item{tool_type}{Tool type ("ko", "pfam", "ec", "hmm", "blast_db")}
}
\value{
Character vector of unique terms
}
\description{
Get detection terms for a tool type
}
