% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato_class.R
\name{get_enzyme_nodes}
\alias{get_enzyme_nodes}
\title{Get enzyme nodes from potato}
\usage{
get_enzyme_nodes(potato)
}
\arguments{
\item{potato}{Potato object}
}
\value{
List of enzyme nodes
}
\description{
Get enzyme nodes from potato
}
