% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato_hash.R
\name{get_potato_hashes}
\alias{get_potato_hashes}
\title{Get potato hashes for multiple potatoes}
\usage{
get_potato_hashes(potatoes)
}
\arguments{
\item{potatoes}{Named list of Potato objects}
}
\value{
Named character vector of hashes
}
\description{
Get potato hashes for multiple potatoes
}
\keyword{internal}
