% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-potato_sack.R
\name{get_tool_version}
\alias{get_tool_version}
\title{Get tool version}
\usage{
get_tool_version(tool_name)
}
\arguments{
\item{tool_name}{Name of the tool (e.g., "exec_annotation", "blastp")}
}
\value{
Version string or "unknown"
}
\description{
Attempts to get version of external tool by running --version
}
\keyword{internal}
