% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-load-sack.R
\name{inspect_saved_sack}
\alias{inspect_saved_sack}
\title{Print metadata about a saved PotatoSack}
\usage{
inspect_saved_sack(path)
}
\arguments{
\item{path}{Path to the RDS file}
}
\value{
Invisibly returns the metadata
}
\description{
Shows information about a saved sack without fully loading it.
}
