% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato_class.R
\name{load_potato}
\alias{load_potato}
\title{Load a potato from JSON}
\usage{
load_potato(path)
}
\arguments{
\item{path}{Path to potato JSON file}
}
\value{
Potato object
}
\description{
Load a potato from JSON
}
