% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato_class.R
\name{load_potatoes}
\alias{load_potatoes}
\title{Load multiple potatoes from directory}
\usage{
load_potatoes(dir, tags = NULL)
}
\arguments{
\item{dir}{Directory containing potato JSON files}

\item{tags}{Optional character vector of tags to filter by}
}
\value{
Named list of Potato objects
}
\description{
Load multiple potatoes from directory
}
