% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-load-sack.R
\name{load_saved_sack}
\alias{load_saved_sack}
\title{Load a PotatoSack object from disk}
\usage{
load_saved_sack(path, validate = TRUE)
}
\arguments{
\item{path}{Path to the RDS file}

\item{validate}{Logical. Validate the loaded object (default TRUE)}
}
\value{
PotatoSack object
}
\description{
Loads a previously saved PotatoSack object from an RDS file.
}
