% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{map_annotation_table}
\alias{map_annotation_table}
\title{Map annotation table to potatoes}
\usage{
map_annotation_table(anno_table, potatoes, config, databases = NULL)
}
\arguments{
\item{anno_table}{Annotation table from initialize_annotation_table()}

\item{potatoes}{List of Potato objects}

\item{config}{Potato config object (to look up database types)}

\item{databases}{Character vector of which databases to map. NULL = all}
}
\value{
Modified annotation table with mapped results
}
\description{
Maps all results in an annotation table to potato genes
}
\keyword{internal}
