% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteCarlo.R
\name{monteCarlo}
\alias{monteCarlo}
\title{monteCarlo}
\usage{
monteCarlo(n)
}
\arguments{
\item{n}{Size of sample}
}
\value{
sample average
}
\description{
computes the sample average from sequence of randomly generated numbers from the uniform distribution
}
