% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_pathway_summary}
\alias{plot_pathway_summary}
\title{Plot pathway completion summary}
\usage{
plot_pathway_summary(sack)
}
\arguments{
\item{sack}{PotatoSack object with scores}
}
\description{
Creates a summary plot showing number of pathways detected per genome.
}
