% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prevalence.R
\name{plot_pathway_uniqueness}
\alias{plot_pathway_uniqueness}
\title{Plot pathway uniqueness distribution}
\usage{
plot_pathway_uniqueness(sack)
}
\arguments{
\item{sack}{PotatoSack object with scoring results}
}
\value{
A ggplot2 object
}
\description{
Creates a histogram showing the distribution of pathway prevalence.
Shows how many pathways are found in 1 genome, 2 genomes, etc.
}
