% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato-package.R
\docType{data}
\name{potato_raster}
\alias{potato_raster}
\title{Geoid potato raster.}
\description{
A simplified world raster of the geoid, obtained from this site in 2014:
http://earth-info.nga.mil/GandG/wgs84/gravitymod/egm2008/GIS/world_geoid
}
