% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato_class.R
\name{print,potato::Potato-method}
\alias{print,potato::Potato-method}
\title{Print method for Potato}
\usage{
## S7 method for class <potato::Potato>
print(x, ...)
}
\description{
Print method for Potato
}
\keyword{internal}
