% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-potato_sack.R
\name{print,potato::PotatoSack-method}
\alias{print,potato::PotatoSack-method}
\title{Print method for PotatoSack}
\usage{
## S7 method for class <potato::PotatoSack>
print(x, ...)
}
\description{
Print method for PotatoSack
}
