% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_table.R
\name{print.potato_annotation_table}
\alias{print.potato_annotation_table}
\title{Print method for annotation table}
\usage{
\method{print}{potato_annotation_table}(x, ...)
}
\description{
Print method for annotation table
}
