% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_provenance.R
\name{print_provenance}
\alias{print_provenance}
\title{Print provenance summary}
\usage{
print_provenance(sack)
}
\arguments{
\item{sack}{PotatoSack object}
}
\description{
Displays provenance information for all annotation and scoring steps
}
