% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato_class.R
\name{print_validation}
\alias{print_validation}
\title{Print validation results}
\usage{
print_validation(validation_result)
}
\arguments{
\item{validation_result}{Result from validate_potato()}
}
\description{
Print validation results
}
