% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_blast.R
\name{run_blast}
\alias{run_blast}
\title{Run BLAST annotation on all genomes}
\usage{
run_blast(
  sack,
  potato_names = NULL,
  conda_env = NULL,
  workers = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{sack}{PotatoSack object}

\item{potato_names}{Vector of potato names (NULL = all)}

\item{conda_env}{Optional conda environment name (defaults to config setting)}

\item{workers}{Number of parallel workers (defaults to config setting, 1 = sequential)}

\item{overwrite}{If FALSE (default), error if blast results already exist}
}
\value{
Modified PotatoSack with blast results in @results$blast
}
\description{
Run BLAST annotation on all genomes
}
