% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hmm.R
\name{run_hmm}
\alias{run_hmm}
\title{Run HMM annotation on all genomes}
\usage{
run_hmm(
  sack,
  potato_names = NULL,
  conda_env = NULL,
  workers = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{sack}{PotatoSack object}

\item{potato_names}{Vector of potato names (NULL = all)}

\item{conda_env}{Optional conda environment name (defaults to config setting)}

\item{workers}{Number of parallel workers (defaults to config setting, 1 = sequential)}

\item{overwrite}{If FALSE (default), error if hmm results already exist}
}
\value{
Modified PotatoSack with hmm results in @results$hmm
}
\description{
Run HMM annotation on all genomes
}
