% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score_single_pathway}
\alias{score_single_pathway}
\title{Score a single genome-potato combination}
\usage{
score_single_pathway(genome_id, potato, hits, min_fraction_override = NULL)
}
\arguments{
\item{genome_id}{Character. Genome identifier}

\item{potato}{Potato S7 object}

\item{hits}{Data frame of annotation hits for this genome (all tools)}

\item{min_fraction_override}{Optional min_fraction override}
}
\value{
List with scoring results
}
\description{
Score a single genome-potato combination
}
\keyword{internal}
