% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_table.R
\name{summarize_annotations}
\alias{summarize_annotations}
\title{Get summary of annotation results}
\usage{
summarize_annotations(anno_table)
}
\arguments{
\item{anno_table}{Annotation table from initialize_annotation_table()}
}
\value{
tibble with summary statistics
}
\description{
Summarizes how many hits each genome has for each tool
}
\keyword{internal}
