% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato_class.R
\name{summary,potato::Potato-method}
\alias{summary,potato::Potato-method}
\title{Summary method for Potato}
\usage{
## S7 method for class <potato::Potato>
summary(object, ...)
}
\description{
Summary method for Potato
}
\keyword{internal}
