% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-potato_sack.R
\name{summary,potato::PotatoSack-method}
\alias{summary,potato::PotatoSack-method}
\title{Summary method for PotatoSack}
\usage{
## S7 method for class <potato::PotatoSack>
summary(object, ...)
}
\description{
Summary method for PotatoSack
}
\keyword{internal}
