% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_table.R
\name{unnest_all_annotations}
\alias{unnest_all_annotations}
\title{Unnest all tool results into long format}
\usage{
unnest_all_annotations(anno_table, tools = NULL)
}
\arguments{
\item{anno_table}{Annotation table}

\item{tools}{Character vector of which tools to include. NULL = all tools}
}
\value{
Long-format data frame with all annotation results
}
\description{
Combines all tool results into a single data frame with tool column
}
\keyword{internal}
