% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_database_configs}
\alias{validate_database_configs}
\title{Validate database configurations}
\usage{
validate_database_configs(databases)
}
\arguments{
\item{databases}{Databases section from config}
}
\description{
Validate database configurations
}
\keyword{internal}
