% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato_class.R
\name{validate_multi_pathway}
\alias{validate_multi_pathway}
\title{Validate multi-pathway network potato (V2 schema)}
\usage{
validate_multi_pathway(data, strict)
}
\description{
Validate multi-pathway network potato (V2 schema)
}
\keyword{internal}
