% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato_class.R
\name{validate_single_pathway}
\alias{validate_single_pathway}
\title{Validate single-pathway potato (legacy schema)}
\usage{
validate_single_pathway(data, strict)
}
\description{
Validate single-pathway potato (legacy schema)
}
\keyword{internal}
