% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato_class.R
\name{build_potato_graph}
\alias{build_potato_graph}
\title{Build igraph from potato}
\usage{
build_potato_graph(potato)
}
\arguments{
\item{potato}{Potato object}
}
\value{
igraph object
}
\description{
Build igraph from potato
}
\keyword{internal}
