% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato_class.R
\name{get_marker_genes}
\alias{get_marker_genes}
\title{Get marker gene nodes from potato}
\usage{
get_marker_genes(potato)
}
\arguments{
\item{potato}{Potato object}
}
\value{
List of nodes marked as marker genes
}
\description{
Get marker gene nodes from potato
}
