% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{map_annotation_table}
\alias{map_annotation_table}
\title{Map annotation table to potatoes}
\usage{
map_annotation_table(anno_table, potatoes, tools = NULL)
}
\arguments{
\item{anno_table}{Annotation table from initialize_annotation_table()}

\item{potatoes}{List of Potato objects}

\item{tools}{Character vector of which tools to map. NULL = all}
}
\value{
Modified annotation table with mapped results
}
\description{
Maps all results in an annotation table to potato genes
}
\keyword{internal}
