% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato_v2.R
\name{plot_v2_interactive}
\alias{plot_v2_interactive}
\title{Interactive visNetwork plot of v2 graph}
\usage{
plot_v2_interactive(g)
}
\arguments{
\item{g}{Graph from build_graph_v2()}
}
\description{
Interactive visNetwork plot of v2 graph
}
