% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato_v2.R
\name{plot_v2_interactive}
\alias{plot_v2_interactive}
\title{Interactive visNetwork plot of v2 graph}
\usage{
plot_v2_interactive(g, layout = "fr")
}
\arguments{
\item{g}{Graph from build_graph_v2()}

\item{layout}{Layout algorithm (e.g., "fr", "kk", "circle", "tree", "grid"). Default: "fr"}
}
\description{
Interactive visNetwork plot of v2 graph
}
