% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato-package.R
\docType{package}
\name{potato-package}
\alias{potato}
\alias{potato-package}
\title{potato: The Pathway annOTATOr}
\description{
POTATO is the R successor to the GATOR (Genome annotATOR) python code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jeffkimbrel.github.io/potato/}
  \item Report bugs at \url{https://github.com/jeffkimbrel/potato/issues}
}

}
\author{
\strong{Maintainer}: Jeff Kimbrel \email{first.last@example.com}

Authors:
\itemize{
  \item Jeff Kimbrel \email{first.last@example.com}
}

}
\keyword{internal}
