% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato-package.R
\docType{package}
\name{potato-package}
\alias{potato}
\alias{potato-package}
\title{potato: Play a Game of 'Potato'}
\description{
Play in your console an interactive version of 'Potato', a one-page role-playing game by Oliver Darkshire.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.github.com/matt-dray/potato}
  \item \url{https://www.patreon.com/deathbybadger}
  \item \url{https://twitter.com/deathbybadger/status/1567425842526945280}
  \item Report bugs at \url{https://www.github.com/matt-dray/potato/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

Other contributors:
\itemize{
  \item Oliver Darkshire [copyright holder]
}

}
\keyword{internal}
