% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato-package.R
\docType{package}
\name{potato-package}
\alias{potato}
\alias{potato-package}
\title{potato: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
