% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potato.R
\name{potato}
\alias{potato}
\title{Start a Game of 'Potato'}
\source{
https://www.patreon.com/deathbybadger
}
\usage{
potato()
}
\value{
Nothing.
}
\description{
Begin an interactive version of 'Potato', a one-page RPG by Oliver Darkshire.
Messages and prompts are printed to the console.
}
\examples{
/dontrun{potato()}
}
